fix: editorをdynamic import に変えた
