+++ src/app/components/editor/editor.tsx
@@ -128,3 +128,4 @@ export const Editor = () => {
     </div>
   );
 };
+
