+++ src/app/components/editor/index.tsx
@@ -1,0 +1,9 @@
+import dynamic from "next/dynamic";
+
+export const Editor = dynamic(
+  () => import("../../components/editor/editor").then((m) => m.Editor),
+  {
+    ssr: false,
+    loading: () => <div>Loading...</div>,
+  }
+);
